services/file/requirements.txt
flask==3.0.2
pyjwt==2.8.0
flask-cors==4.0.0
